pyproject.toml
[tool.poetry]
name = "command-line content feed"
version = "0.1.0"
description = "Feed for various content sources, all in the command line"
authors = ["Andres Ponce <andreseeponce@gmail.com>"]
license = "GPL v2"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
urwid = {extras = ["curses"], version = "^2.6.15"}
pyyaml = "^6.0.2"
pre-commit = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
